refactor(utils): extract shared rule builders for zod plugins
