+++ packages/utils/src/rule-builders/consistent-object-schema-type-create.ts
@@ -1,0 +1,74 @@
+import type { TSESLint } from '@typescript-eslint/utils';
+import { AST_NODE_TYPES } from '@typescript-eslint/utils';
+
+import { createZodSchemaImportTrack } from '../track-zod-schema-imports.js';
+import type { ZodImportScope } from '../zod-import-scope.js';
+
+const ZOD_OBJECT_METHODS = ['object', 'looseObject', 'strictObject'] as const;
+
+type ZodObjectMethod = (typeof ZOD_OBJECT_METHODS)[number];
+
+interface Options {
+  allow: Array<ZodObjectMethod>;
+}
+
+type MessageIds = 'consistentMethod' | 'useMethod';
+
+export function buildConsistentObjectSchemaTypeCreate(
+  scope: ZodImportScope,
+): (
+  context: Readonly<TSESLint.RuleContext<MessageIds, [Options]>>,
+  options: readonly [Options],
+) => TSESLint.RuleListener {
+  const { trackZodSchemaImports } = createZodSchemaImportTrack(scope);
+
+  return function create(context, [{ allow: allowedList }]) {
+    const { importDeclarationListener, detectZodSchemaRootNode } = trackZodSchemaImports();
+
+    return {
+      ImportDeclaration: importDeclarationListener,
+      CallExpression(node): void {
+        const zodSchemaMeta = detectZodSchemaRootNode(node);
+
+        const schemaType = zodSchemaMeta?.schemaType as ZodObjectMethod | undefined;
+
+        if (!schemaType || !ZOD_OBJECT_METHODS.includes(schemaType)) {
+          return;
+        }
+
+        if (allowedList.includes(schemaType)) {
+          return;
+        }
+
+        const { callee } = node;
+
+        if (callee.type === AST_NODE_TYPES.Identifier) {
+          context.report({
+            node,
+            messageId: 'consistentMethod',
+            data: { actual: schemaType, allowedList: allowedList.join(',') },
+          });
+          return;
+        }
+
+        if (callee.type === AST_NODE_TYPES.MemberExpression) {
+          context.report({
+            node,
+            messageId: 'consistentMethod',
+            data: {
+              actual: schemaType,
+              allowedList: allowedList.join(','),
+            },
+            suggest: allowedList.map<TSESLint.ReportSuggestionArray<MessageIds>[number]>((it) => ({
+              messageId: 'useMethod',
+              data: { expected: it },
+              fix(fixer): TSESLint.RuleFix {
+                return fixer.replaceText(callee.property, it);
+              },
+            })),
+          });
+        }
+      },
+    };
+  };
+}
