+++ packages/utils/src/rule-builders/consistent-schema-var-name-create.ts
@@ -1,0 +1,65 @@
+import type { TSESLint } from '@typescript-eslint/utils';
+import { AST_NODE_TYPES } from '@typescript-eslint/utils';
+
+import { createZodSchemaImportTrack } from '../track-zod-schema-imports.js';
+import type { ZodImportScope } from '../zod-import-scope.js';
+import { ZOD_NON_SCHEMA_PRODUCING_METHODS } from '../zod-non-schema-producing-methods.js';
+
+interface Options {
+  before?: string;
+  after?: string;
+}
+
+export function buildConsistentSchemaVarNameCreate(
+  scope: ZodImportScope,
+): (
+  context: Readonly<TSESLint.RuleContext<'invalidName', [Options]>>,
+  options: readonly [Options],
+) => TSESLint.RuleListener {
+  const { trackZodSchemaImports } = createZodSchemaImportTrack(scope);
+
+  return function create(context, [{ before = '', after = '' }]) {
+    const { importDeclarationListener, detectZodSchemaRootNode, collectZodChainMethods } =
+      trackZodSchemaImports();
+
+    return {
+      ImportDeclaration: importDeclarationListener,
+      VariableDeclarator(node): void {
+        const initNode = node.init;
+
+        if (
+          initNode?.type !== AST_NODE_TYPES.CallExpression ||
+          !detectZodSchemaRootNode(initNode)
+        ) {
+          return;
+        }
+
+        const chainMethods = collectZodChainMethods(initNode).map((it) => it.name);
+
+        if (ZOD_NON_SCHEMA_PRODUCING_METHODS.some((it) => chainMethods.includes(it))) {
+          return;
+        }
+
+        if (node.id.type !== AST_NODE_TYPES.Identifier) {
+          return;
+        }
+
+        const { name } = node.id;
+        const validPrefix = !before || name.startsWith(before);
+        const validSuffix = !after || name.endsWith(after);
+
+        if (validPrefix && validSuffix) {
+          return;
+        }
+
+        const expected = (validPrefix ? '' : before) + name + (validSuffix ? '' : after);
+
+        context.report({
+          node,
+          messageId: 'invalidName',
+          data: { expected },
+        });
+      },
+    };
+  };
+}
