+++ packages/utils/src/rule-builders/require-brand-type-parameter-create.ts
@@ -1,0 +1,58 @@
+import type { TSESLint, TSESTree } from '@typescript-eslint/utils';
+
+import { createZodSchemaImportTrack } from '../track-zod-schema-imports.js';
+import type { ZodImportScope } from '../zod-import-scope.js';
+
+type MessageIds = 'missingTypeParameter' | 'removeBrandFunction';
+
+export function buildRequireBrandTypeParameterCreate(
+  scope: ZodImportScope,
+): (context: Readonly<TSESLint.RuleContext<MessageIds, []>>) => TSESLint.RuleListener {
+  const { trackZodSchemaImports } = createZodSchemaImportTrack(scope);
+
+  return function create(context) {
+    const { importDeclarationListener, detectZodSchemaRootNode, collectZodChainMethods } =
+      trackZodSchemaImports();
+
+    return {
+      ImportDeclaration: importDeclarationListener,
+      CallExpression(node): void {
+        const zodSchemaMeta = detectZodSchemaRootNode(node);
+        if (!zodSchemaMeta) {
+          return;
+        }
+
+        const methods = collectZodChainMethods(zodSchemaMeta.node);
+
+        const brandMethod = methods.find((it) => it.name === 'brand');
+
+        if (!brandMethod) {
+          return;
+        }
+
+        const brandNode = brandMethod.node;
+
+        const { typeArguments } = brandNode;
+
+        if (typeArguments && typeArguments.params.length > 0) {
+          return;
+        }
+
+        const brandCalleeNode = brandNode.callee as TSESTree.MemberExpression;
+
+        context.report({
+          messageId: 'missingTypeParameter',
+          node: brandCalleeNode.property,
+          suggest: [
+            {
+              messageId: 'removeBrandFunction',
+              fix(fixer): TSESLint.RuleFix {
+                return fixer.removeRange([brandCalleeNode.object.range[1], brandNode.range[1]]);
+              },
+            },
+          ],
+        });
+      },
+    };
+  };
+}
