+++ plugins/eslint-plugin-zod-mini/src/rules/require-brand-type-parameter.ts
@@ -1,9 +1,6 @@
-import { createZodSchemaImportTrack, zodMiniImportScope } from '@eslint-zod/utils';
-import type { TSESLint, TSESTree } from '@typescript-eslint/utils';
+import { buildRequireBrandTypeParameterCreate, zodMiniImportScope } from '@eslint-zod/utils';
 
 import { createZodMiniPluginRule } from '../utils/create-plugin-rule.js';
-
-const { trackZodSchemaImports } = createZodSchemaImportTrack(zodMiniImportScope);
 
 export const requireBrandTypeParameter = createZodMiniPluginRule({
   name: 'require-brand-type-parameter',
@@ -21,49 +18,5 @@ export const requireBrandTypeParameter = createZodMiniPluginRule({
     schema: [],
   },
   defaultOptions: [],
-  create(context) {
-    const { importDeclarationListener, detectZodSchemaRootNode, collectZodChainMethods } =
-      trackZodSchemaImports();
-
-    return {
-      ImportDeclaration: importDeclarationListener,
-      CallExpression(node): void {
-        const zodSchemaMeta = detectZodSchemaRootNode(node);
-        if (!zodSchemaMeta) {
-          return;
-        }
-
-        const methods = collectZodChainMethods(zodSchemaMeta.node);
-
-        const brandMethod = methods.find((it) => it.name === 'brand');
-
-        if (!brandMethod) {
-          return;
-        }
-
-        const brandNode = brandMethod.node;
-
-        const { typeArguments } = brandNode;
-
-        if (typeArguments && typeArguments.params.length > 0) {
-          return;
-        }
-
-        const brandCalleeNode = brandNode.callee as TSESTree.MemberExpression;
-
-        context.report({
-          messageId: 'missingTypeParameter',
-          node: brandCalleeNode.property,
-          suggest: [
-            {
-              messageId: 'removeBrandFunction',
-              fix(fixer): TSESLint.RuleFix {
-                return fixer.removeRange([brandCalleeNode.object.range[1], brandNode.range[1]]);
-              },
-            },
-          ],
-        });
-      },
-    };
-  },
+  create: buildRequireBrandTypeParameterCreate(zodMiniImportScope),
 });
