+++ plugins/eslint-plugin-zod-mini/src/rules/schema-error-property-style.ts
@@ -1,13 +1,8 @@
-import { createZodSchemaImportTrack, zodMiniImportScope } from '@eslint-zod/utils';
-import type { TSESTree } from '@typescript-eslint/utils';
-import { AST_NODE_TYPES } from '@typescript-eslint/utils';
-import esquery from 'esquery';
+import { buildSchemaErrorPropertyStyleCreate, zodMiniImportScope } from '@eslint-zod/utils';
 
 import { createZodMiniPluginRule } from '../utils/create-plugin-rule.js';
 
 type MessageIds = 'invalidStyle' | 'invalidSelector';
-
-const { trackZodSchemaImports } = createZodSchemaImportTrack(zodMiniImportScope);
 
 export const schemaErrorPropertyStyle = createZodMiniPluginRule<
   [{ selector: string; example: string }],
@@ -43,96 +38,5 @@ export const schemaErrorPropertyStyle = createZodMiniPluginRule<
     ],
   },
   defaultOptions: [{ selector: 'Literal,TemplateLiteral', example: "'error message'" }],
-  create(context, [{ selector, example }]) {
-    const { importDeclarationListener, detectZodSchemaRootNode, collectZodChainMethods } =
-      trackZodSchemaImports();
-
-    /**
-     * Parsing `selector` to ensure it is valid,
-     * if not report an error and return empty rule listener
-     */
-    try {
-      esquery.parse(selector);
-    } catch {
-      context.report({
-        loc: { line: 1, column: 0 },
-        messageId: 'invalidSelector',
-        data: { selector },
-      });
-      return {};
-    }
-
-    return {
-      ImportDeclaration: importDeclarationListener,
-      CallExpression(node): void {
-        const zodSchemaMeta = detectZodSchemaRootNode(node);
-        if (!zodSchemaMeta) {
-          return;
-        }
-
-        if (
-          zodSchemaMeta.schemaType !== 'custom' &&
-          !collectZodChainMethods(node).some((it) => it.name === 'refine')
-        ) {
-          return;
-        }
-
-        // Error should be the second parameter,
-        // if not present stop processing
-        if (node.arguments.length < 2) {
-          return;
-        }
-
-        let errorMessageNode: TSESTree.Node | undefined;
-
-        const [, params] = node.arguments;
-
-        switch (params.type) {
-          case AST_NODE_TYPES.Literal:
-          case AST_NODE_TYPES.TemplateLiteral:
-            errorMessageNode = params;
-            break;
-
-          case AST_NODE_TYPES.ObjectExpression:
-            for (const property of params.properties) {
-              if (
-                property.type === AST_NODE_TYPES.Property &&
-                property.key.type === AST_NODE_TYPES.Identifier &&
-                property.key.name === 'error'
-              ) {
-                errorMessageNode = property.value;
-                break;
-              }
-            }
-            break;
-
-          // no default
-        }
-
-        if (!errorMessageNode) {
-          return;
-        }
-
-        const match = esquery.matches(
-          errorMessageNode as never,
-          esquery.parse(selector),
-          errorMessageNode as never,
-        );
-
-        if (match) {
-          return;
-        }
-
-        context.report({
-          node,
-          messageId: 'invalidStyle',
-          data: {
-            selector,
-            example,
-            actual: context.sourceCode.getText(errorMessageNode),
-          },
-        });
-      },
-    };
-  },
+  create: buildSchemaErrorPropertyStyleCreate(zodMiniImportScope),
 });
