+++ plugins/eslint-plugin-zod/src/rules/consistent-schema-output-type-style.ts
@@ -1,6 +1,4 @@
-import { createZodSchemaImportTrack, zodImportScope } from '@eslint-zod/utils';
-import type { TSESTree } from '@typescript-eslint/utils';
-import { AST_NODE_TYPES } from '@typescript-eslint/utils';
+import { buildConsistentSchemaOutputTypeStyleCreate, zodImportScope } from '@eslint-zod/utils';
 
 import { createZodPluginRule } from '../utils/create-plugin-rule.js';
 
@@ -13,8 +11,6 @@ interface Options {
 }
 
 type MessageIds = 'useInfer' | 'useOutput';
-
-const { trackZodSchemaImports } = createZodSchemaImportTrack(zodImportScope);
 
 export const consistentSchemaOutputTypeStyle = createZodPluginRule<[Options], MessageIds>({
   name: 'consistent-schema-output-type-style',
@@ -43,67 +39,5 @@ export const consistentSchemaOutputTypeStyle = createZodPluginRule<[Options], Me
     ],
   },
   defaultOptions: [{ style: 'output' }],
-  create(context, [{ style }]) {
-    const {
-      importDeclarationListener,
-      isZodNamespace,
-      getNamedImportOriginal,
-      getNamedImportLocal,
-    } = trackZodSchemaImports();
-
-    return {
-      ImportDeclaration: importDeclarationListener,
-
-      TSTypeReference(node: TSESTree.TSTypeReference): void {
-        const { typeName } = node;
-
-        if (typeName.type === AST_NODE_TYPES.TSQualifiedName) {
-          const { left, right } = typeName;
-
-          if (left.type !== AST_NODE_TYPES.Identifier || !isZodNamespace(left.name)) {
-            return;
-          }
-
-          const usedStyle = right.name;
-
-          if ((usedStyle !== 'infer' && usedStyle !== 'output') || usedStyle === style) {
-            return;
-          }
-
-          context.report({
-            node: right,
-            messageId: style === 'infer' ? 'useInfer' : 'useOutput',
-            fix(fixer) {
-              return fixer.replaceText(right, style);
-            },
-          });
-
-          return;
-        }
-
-        if (typeName.type === AST_NODE_TYPES.Identifier) {
-          const originalName = getNamedImportOriginal(typeName.name);
-
-          if ((originalName !== 'infer' && originalName !== 'output') || originalName === style) {
-            return;
-          }
-
-          const targetLocalName = getNamedImportLocal(style);
-
-          context.report({
-            node: typeName,
-            messageId: style === 'infer' ? 'useInfer' : 'useOutput',
-            fix(fixer) {
-              // 'infer' cannot be used as a standalone type name (TypeScript keyword),
-              // so only fix when the target local name is safe to use directly.
-              if (!targetLocalName || targetLocalName === 'infer') {
-                return null;
-              }
-              return fixer.replaceText(typeName, targetLocalName);
-            },
-          });
-        }
-      },
-    };
-  },
+  create: buildConsistentSchemaOutputTypeStyleCreate(zodImportScope),
 });
