+++ plugins/eslint-plugin-zod/src/rules/no-any-schema.ts
@@ -1,10 +1,6 @@
-import { createZodSchemaImportTrack, zodImportScope } from '@eslint-zod/utils';
-import type { TSESLint } from '@typescript-eslint/utils';
-import { AST_NODE_TYPES } from '@typescript-eslint/utils';
+import { buildNoAnySchemaCreate, zodImportScope } from '@eslint-zod/utils';
 
 import { createZodPluginRule } from '../utils/create-plugin-rule.js';
-
-const { trackZodSchemaImports } = createZodSchemaImportTrack(zodImportScope);
 
 export const noAnySchema = createZodPluginRule({
   name: 'no-any-schema',
@@ -21,59 +17,5 @@ export const noAnySchema = createZodPluginRule({
     schema: [],
   },
   defaultOptions: [],
-  create(context) {
-    const { importDeclarationListener, detectZodSchemaRootNode, collectZodChainMethods } =
-      trackZodSchemaImports();
-
-    return {
-      ImportDeclaration: importDeclarationListener,
-      CallExpression(node): void {
-        const zodSchemaMeta = detectZodSchemaRootNode(node);
-
-        if (zodSchemaMeta?.schemaType !== 'any') {
-          return;
-        }
-
-        const { callee } = node;
-
-        if (callee.type === AST_NODE_TYPES.Identifier) {
-          context.report({
-            node,
-            messageId: 'noZAny',
-          });
-          return;
-        }
-
-        if (callee.type === AST_NODE_TYPES.MemberExpression) {
-          const [{ node: schemaMethod }] = collectZodChainMethods(node);
-
-          const schemaMethodCallee = schemaMethod.callee;
-
-          if (
-            schemaMethodCallee.type === AST_NODE_TYPES.MemberExpression &&
-            schemaMethodCallee.property.type === AST_NODE_TYPES.Identifier
-          ) {
-            context.report({
-              node,
-              messageId: 'noZAny',
-              suggest: [
-                {
-                  messageId: 'useUnknown',
-                  fix(fixer): TSESLint.RuleFix {
-                    return fixer.replaceText(schemaMethodCallee.property, 'unknown');
-                  },
-                },
-              ],
-            });
-            return;
-          }
-
-          context.report({
-            node,
-            messageId: 'noZAny',
-          });
-        }
-      },
-    };
-  },
+  create: buildNoAnySchemaCreate(zodImportScope),
 });
