+++ plugins/eslint-plugin-zod/src/rules/no-empty-custom-schema.ts
@@ -1,8 +1,6 @@
-import { createZodSchemaImportTrack, zodImportScope } from '@eslint-zod/utils';
+import { buildNoEmptyCustomSchemaCreate, zodImportScope } from '@eslint-zod/utils';
 
 import { createZodPluginRule } from '../utils/create-plugin-rule.js';
-
-const { trackZodSchemaImports } = createZodSchemaImportTrack(zodImportScope);
 
 export const noEmptyCustomSchema = createZodPluginRule({
   name: 'no-empty-custom-schema',
@@ -18,37 +16,5 @@ export const noEmptyCustomSchema = createZodPluginRule({
     schema: [],
   },
   defaultOptions: [],
-  create(context) {
-    const {
-      //
-      importDeclarationListener,
-      detectZodSchemaRootNode,
-      collectZodChainMethods,
-    } = trackZodSchemaImports();
-
-    return {
-      ImportDeclaration: importDeclarationListener,
-      CallExpression(node): void {
-        const zodSchemaMeta = detectZodSchemaRootNode(node);
-        if (!zodSchemaMeta) {
-          return;
-        }
-
-        if (zodSchemaMeta.schemaType !== 'custom') {
-          return;
-        }
-
-        // Find the actual custom() call node in the chain
-        const chainMethods = collectZodChainMethods(node);
-        const customCallNode = chainMethods.find((method) => method.name === 'custom')?.node;
-
-        if (customCallNode?.arguments.length === 0) {
-          context.report({
-            node: customCallNode,
-            messageId: 'noEmptyCustomSchema',
-          });
-        }
-      },
-    };
-  },
+  create: buildNoEmptyCustomSchemaCreate(zodImportScope),
 });
