+++ plugins/eslint-plugin-zod/src/rules/no-unknown-schema.ts
@@ -1,8 +1,6 @@
-import { createZodSchemaImportTrack, zodImportScope } from '@eslint-zod/utils';
+import { buildNoUnknownSchemaCreate, zodImportScope } from '@eslint-zod/utils';
 
 import { createZodPluginRule } from '../utils/create-plugin-rule.js';
-
-const { trackZodSchemaImports } = createZodSchemaImportTrack(zodImportScope);
 
 export const noUnknownSchema = createZodPluginRule({
   name: 'no-unknown-schema',
@@ -17,25 +15,5 @@ export const noUnknownSchema = createZodPluginRule({
     schema: [],
   },
   defaultOptions: [],
-  create(context) {
-    const {
-      //
-      importDeclarationListener,
-      detectZodSchemaRootNode,
-    } = trackZodSchemaImports();
-
-    return {
-      ImportDeclaration: importDeclarationListener,
-      CallExpression(node): void {
-        const zodSchemaMeta = detectZodSchemaRootNode(node);
-
-        if (zodSchemaMeta?.schemaType === 'unknown') {
-          context.report({
-            node,
-            messageId: 'noZUnknown',
-          });
-        }
-      },
-    };
-  },
+  create: buildNoUnknownSchemaCreate(zodImportScope),
 });
